关卡流程添加 胜利之后的刷新。角色属性读表，角色距离去除小数点后4位防止 斜着的格子距离因为计算结果精度 大于1。属性全部改为对象池

+++ Client/Assets/Scripts/SkillSystem/Commands/CommandCauseDamage.cs
@@ -47,7 +47,7 @@ namespace SkillSystem
                 {
                     if (CurDamageComputeType == DamageComputeType.NormalDamage)
                     {
-                        damageParam = (float)caster.GetAttribute(AttributeType.AttackDamage).GetFinalValue();
+                        damageParam = (int)caster.GetAttribute(AttributeType.AttackDamage).GetFinalValue();
                     }
                     else if (CurDamageComputeType == DamageComputeType.FixNumAddAttrPercent)
                     {
